+++ src/handlers/home.rs
@@ -2,10 +2,10 @@ use crate::cli::serve::AppState;
 use crate::entities::page::Page;
 use crate::entities::pagination::Pagination;
 use crate::utils::file::read_file;
-use crate::utils::template::minify_html;
 use axum::extract::{Query, State};
 use axum::response::Html;
 use glob::glob;
+use minify::html::minify;
 use std::sync::Arc;
 use tera::Context;
 
@@ -82,5 +82,5 @@ pub async fn show(
     context.insert("pages", &pages);
 
     let html = state.tera.render("home.html", &context).unwrap();
-    Html(minify_html(html))
+    Html(minify(&html))
 }