+++ src/utils.rs
@@ -1,3 +1,2 @@
 pub mod file;
 pub mod handler;
-pub mod template;